added some changes related to the comments, also changes some names of the packages
