+++ src/main/java/homeworks/christian/firstAssigment/Human.java
@@ -1,4 +1,4 @@
-package homeworks.christian.firstAssigment.firstAssigment;
+package homeworks.christian.firstAssigment;
 
 public class Human implements Interface{
