+++ src/main/java/homeworks/christian/firstAssigment/Interface.java
@@ -1,4 +1,4 @@
-package homeworks.christian.firstAssigment.firstAssigment;
+package homeworks.christian.firstAssigment;
 
 public interface Interface {
