+++ src/main/java/homeworks/christian/firstAssigment/Woman.java
@@ -1,4 +1,4 @@
-package homeworks.christian.firstAssigment.firstAssigment;
+package homeworks.christian.firstAssigment;
 
 public class Woman extends Human {
