+++ src/main/java/homeworks/christian/secondAssigment/ClassA.java
@@ -1,43 +1,26 @@
-package homeworks.christian.firstAssigment.secondAssigment;
+package homeworks.christian.secondAssigment;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class ClassE implements InterfaceCEG {
+public class ClassA extends ParentABD{
 
-    private int h;
     private List<Integer> c = new ArrayList<Integer>();
 
     /**
      * this method is empty and is a override
      */
     @Override
-    public void methodD() {
-
+    public void MethodB() {
+        super.MethodB();
     }
 
     /**
      * this method is empty and is a override
      */
     @Override
-    public void methodE() {
-
-    }
-
-    /**
-     * method get h
-     * @return h
-     */
-    public int getH() {
-        return h;
-    }
-
-    /**
-     * method set h
-     * @param h
-     */
-    public void setH(int h) {
-        this.h = h;
+    public void MethodC() {
+        super.MethodC();
     }
 
     /**
@@ -55,4 +38,5 @@ public class ClassE implements InterfaceCEG {
     public void setC(List<Integer> c) {
         this.c = c;
     }
+
 }
