+++ src/main/java/homeworks/christian/secondAssigment/ClassC.java
@@ -1,4 +1,4 @@
-package homeworks.christian.firstAssigment.secondAssigment;
+package homeworks.christian.secondAssigment;
 
 public class ClassC implements InterfaceCEG{
