+++ src/main/java/homeworks/christian/secondAssigment/ParentABD.java
@@ -1,4 +1,4 @@
-package homeworks.christian.firstAssigment.secondAssigment;
+package homeworks.christian.secondAssigment;
 
 public class ParentABD {
